Refactor the static config + utils files to read in the new format of input files.
