+++ test/test_match_tournament_generation.py
@@ -1,9 +1,11 @@
 import shutil
 
 import scripts.helpers.helper_utils as helper
-import utils.match_utils.data_reader_utils as reader
 import pytest
 import os
+
+import utils.match_utils.matches
+import utils.match_utils.tournaments
 
 
 @pytest.fixture
@@ -17,11 +19,10 @@ def correct_file_dir():
 
 def test_correct_files(correct_file_dir):
     assert (os.path.isdir(correct_file_dir))
-    tournament = reader.Tournaments("{}/tournaments.csv".format(correct_file_dir))
-    matches = reader.Matches("{}/tournaments.csv".format(correct_file_dir))
+    tournament = utils.match_utils.tournaments.Tournaments("{}/tournaments.csv".format(correct_file_dir))
     print(tournament.df.columns.values)
 
-    expected_columns =  ["tournament", "start_date", "end_date"]
+    expected_columns = ["tournament", "start_date", "end_date"]
     for column in expected_columns:
         assert(column in tournament.df.columns.values)
