+++ utils/match_utils/players.py
@@ -1,0 +1,6 @@
+import pandas as pd
+
+
+class Players:
+    def __init__(self, player_file):
+        self.players_df = pd.read_csv(player_file)
